% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetchTimeSeries}
\alias{fetchTimeSeries}
\title{Fetch time series data}
\usage{
fetchTimeSeries(timeType = "intraday", symbol = "spy", interval = "5min",
  outputsize = "full", datatype = "csv")
}
\arguments{
\item{timeType}{This parameter specifies the time series type. Valid character string input: \code{intraday}, \code{daily}, \code{daily_adjusted}, \code{weekly}, \code{monthly}.}

\item{symbol}{Stock symbol, such as MSFT for Microsoft.}

\item{interval}{Required argument for intraday trading. Valid character string input: \code{1min}, \code{5min}, \code{15min}, \code{60min}.}

\item{outputsize}{Either \code{compact} or \code{full}. Returns only 100 data points versus full-length time series.}

\item{datatype}{Either \code{csv} or \code{json}.}
}
\value{
An \code{xts} object with the time series data.
}
\description{
This function allows the user to pull data of the five different time series types offered from Alpha Vantage: Intraday, Daily, Daily (Adj. Close),
Weekly, and Monthly.
}
