% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlphaVantageClient.R
\name{setAPIKey}
\alias{setAPIKey}
\title{Assign API key to global environment.}
\usage{
setAPIKey(key)
}
\arguments{
\item{key}{The API key from AlphaVantage}
}
\description{
Assign API key to global environment.
}
