% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{technical_indicator_specification}
\alias{technical_indicator_specification}
\title{Required and optional arguments specification.}
\format{An object of class \code{list} of length 54.}
\usage{
technical_indicator_specification
}
\description{
A listing that replicates the required and optional argument listings from the Alpha Vantage documentation.
}
\keyword{datasets}
