% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{validation_list}
\alias{validation_list}
\title{Validation listings}
\format{An object of class \code{list} of length 5.}
\usage{
validation_list
}
\description{
A listing that outlines which arguments are acceptable.
}
\keyword{datasets}
